llm-service/requirements.txt
flask==3.0.2
sentence-transformers==2.3.1
chromadb==0.4.24
trino==0.329
python-dotenv==1.0.1
ollama==0.1.3
openai==1.12.0 
pandas==2.2.0
numpy==1.26.3
pyarrow==14.0.2
faker==22.6.0